requirements.txt
# Top-level dev dependencies
cryptography
google-api-python-client
google-auth
google-auth-oauthlib
mypy
pyyaml
types-PyYAML
